Added update function
I added the update function which has similar functions to create repo but it checks if the repo path exists already and outputs error to the log if it does not exist.

Added an else if block in the responsePage for the error if the repo path does not exist.